# openapi.yaml
openapi: 3.1.0
info:
  title: Aurite Agents API
  description: |
    API for the Aurite Agents framework - a Python framework for building AI agents using the Model Context Protocol (MCP).

    ## Authentication
    All endpoints (except /health) require an API key to be passed in the `X-API-Key` header.

    ## Base URL
    The default base URL is `http://localhost:8000`
  version: 1.0.0
  contact:
    name: Aurite Agents Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - APIKeyHeader: []
paths:
...
  /config/components:
    get:
      summary: List Component Types
      description: List all available component types.
      tags:
        - Configuration Manager
      responses:
        "200":
          description: List of component types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  [
                    "agent",
                    "llm",
                    "mcp_server",
                    "linear_workflow",
                    "custom_workflow",
                  ]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /config/components/{component_type}:
    get:
      summary: List Components By Type
      description: |
        List all available components of a specific type.
        Accepts both singular and plural forms (e.g., 'agent' or 'agents').
      tags:
        - Configuration Manager
      parameters:
        - name: component_type
          in: path
          required: true
          schema:
            type: string
            enum:
              [
                agent,
                agents,
                llm,
                llms,
                mcp_server,
                mcp_servers,
                linear_workflow,
                linear_workflows,
                custom_workflow,
                custom_workflows,
              ]
          description: Type of component
      responses:
        "200":
          description: List of components
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

    post:
      summary: Create Component
      description: Create a new component of the specified type.
      tags:
        - Configuration Manager
      parameters:
        - name: component_type
          in: path
          required: true
          schema:
            type: string
            enum:
              [
                agent,
                agents,
                llm,
                llms,
                mcp_server,
                mcp_servers,
                linear_workflow,
                linear_workflows,
                custom_workflow,
                custom_workflows,
              ]
          description: Type of component
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentCreate"
      responses:
        "200":
          description: Component created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Component already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /config/components/{component_type}/{component_id}:
    get:
      summary: Get Component By Id
      description: |
        Get a specific component by its type and ID.
        Accepts both singular and plural forms for component type (e.g., 'agent' or 'agents').
      tags:
        - Configuration Manager
      parameters:
        - name: component_type
          in: path
          required: true
          schema:
            type: string
          description: Type of component
        - name: component_id
          in: path
          required: true
          schema:
            type: string
          description: Component ID or name
      responses:
        "200":
          description: Component configuration
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    put:
      summary: Update Component
      description: Update an existing component.
      tags:
        - Configuration Manager
      parameters:
        - name: component_type
          in: path
          required: true
          schema:
            type: string
          description: Type of component
        - name: component_id
          in: path
          required: true
          schema:
            type: string
          description: Component ID or name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentUpdate"
      responses:
        "200":
          description: Component updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete Component
      description: Delete a component.
      tags:
        - Configuration Manager
      parameters:
        - name: component_type
          in: path
          required: true
          schema:
            type: string
          description: Type of component
        - name: component_id
          in: path
          required: true
          schema:
            type: string
          description: Component ID or name
      responses:
        "200":
          description: Component deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /config/components/{component_type}/{component_id}/validate:
    post:
      summary: Validate Component
      description: Validate a component's configuration.
      tags:
        - Configuration Manager
      parameters:
        - name: component_type
          in: path
          required: true
          schema:
            type: string
          description: Type of component
        - name: component_id
          in: path
          required: true
          schema:
            type: string
          description: Component ID or name
      responses:
        "200":
          description: Component is valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /config/sources:
    get:
      summary: List Config Sources
      description: |
        List all configuration source directories with context information.

        Returns a list of configuration source directories in priority order,
        including their context (project/workspace/user) and associated names.
      tags:
        - Configuration Manager
      responses:
        "200":
          description: List of configuration sources
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /config/files/{source_name}:
    get:
      summary: List Config Files By Source
      description: |
        List all configuration files for a specific source.

        Returns a list of relative file paths for the given source.
      tags:
        - Configuration Manager
      parameters:
        - name: source_name
          in: path
          required: true
          schema:
            type: string
          description: Name of the configuration source
      responses:
        "200":
          description: List of configuration files
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /config/files/{source_name}/{file_path}:
    get:
      summary: Get File Content
      description: Get the content of a specific configuration file.
      tags:
        - Configuration Manager
      parameters:
        - name: source_name
          in: path
          required: true
          schema:
            type: string
          description: Name of the configuration source
        - name: file_path
          in: path
          required: true
          schema:
            type: string
          description: Path to the configuration file
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    put:
      summary: Update Config File
      description: Update an existing configuration file.
      tags:
        - Configuration Manager
      parameters:
        - name: source_name
          in: path
          required: true
          schema:
            type: string
          description: Name of the configuration source
        - name: file_path
          in: path
          required: true
          schema:
            type: string
          description: Path to the configuration file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdateRequest"
      responses:
        "200":
          description: File updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      summary: Delete Config File
      description: Delete an existing configuration file.
      tags:
        - Configuration Manager
      parameters:
        - name: source_name
          in: path
          required: true
          schema:
            type: string
          description: Name of the configuration source
        - name: file_path
          in: path
          required: true
          schema:
            type: string
          description: Path to the configuration file
      responses:
        "200":
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /config/files:
    post:
      summary: Create Config File
      description: Create a new configuration file.
      tags:
        - Configuration Manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileCreateRequest"
      responses:
        "200":
          description: File created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: File already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /config/validate:
    post:
      summary: Validate All Components
      description: Validate all components in the system.
      tags:
        - Configuration Manager
      responses:
        "200":
          description: Validation results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
  /config/refresh:
    post:
      summary: Refresh Configs
      description: |
        Force refresh all configurations from disk.
        This will reload all configuration files and rebuild the component index.
      tags:
        - Configuration Manager
      responses:
        "200":
          description: Configurations refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Failed to refresh configurations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /config/projects/active:
    get:
      summary: Get Active Project
      description: Get information about the currently active project.
      tags:
        - Configuration Manager
      responses:
        "200":
          description: Active project information
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/ProjectInfo"
                  - type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /config/workspaces:
    get:
      summary: List Workspaces
      description: List workspace information (currently supports single workspace).
      tags:
        - Configuration Manager
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      description: Standard message response

    # MCP Host Schemas
...
    ComponentCreate:
      type: object
      required:
        - name
        - config
      properties:
        name:
          type: string
          description: Unique name for the component
        config:
          type: object
          additionalProperties: true
          description: Component configuration
      description: Request model for creating a new component
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
...